requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-RESTX==1.2.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # For PostgreSQL
mysqlclient==2.2.4  # For MySQL

# Async & Background
celery==5.3.4
redis==5.0.1

# Security & Validation
Werkzeug==3.0.1
email-validator==2.1.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.2

# Monitoring & Logging
prometheus-client==0.19.0
blinker==1.7.0
structlog==23.2.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
geoip2==4.8.0  # For IP geolocation

# Development
pytest==7.4.3
pytest-flask==1.2.0
pytest-mock==3.11.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0